Compute RoPE base_freqs in float32
The logspace->linspace+exp substitution addresses the missing
MPS logspace kernel. Computing it in float32 and casting
the result to compute_dtype (e.g. bf16 under mixed precision)
matches the reference implementation and improves relative error
vs torch.logspace and jax.numpy.logspace.

+++ src/alphagenome_pytorch/attention.py
@@ -66,15 +66,14 @@ def apply_rope(x, positions=None, max_position=_MAX_RELATIVE_DISTANCE, inplace=F
     num_freq = C // 2
     # JAX geomspace equivalent: geomspace(1, max_position - num_freq + 1, num_freq).
     # Use torch.exp(linspace * log(base)) instead of torch.logspace so this stays
-    # on-device on MPS, which lacks an aten::logspace.out kernel
-    # (pytorch/pytorch#141287). Mathematically identical; bit-equivalent within
-    # float32 ulp (max abs diff 7e-7 verified on CPU).
+    # on-device on MPS, which lacks an aten::logspace.out kernel (pytorch/pytorch#141287).
+    # Compute in float32 and cast the sum (matching JAX's .astype).
     log_end = math.log10(max_position - num_freq + 1)
     base_freqs = torch.exp(
         torch.linspace(0.0, log_end, steps=num_freq,
-                       device=x.device, dtype=compute_dtype) * math.log(10.0)
+                       device=x.device, dtype=torch.float32) * math.log(10.0)
     )
-    denom = torch.arange(num_freq, device=x.device, dtype=compute_dtype) + base_freqs
+    denom = (torch.arange(num_freq, device=x.device, dtype=torch.float32) + base_freqs).to(compute_dtype)
     inv_freq = 1.0 / denom
 
     theta = torch.einsum('bs,f->bsf', positions, inv_freq)
